requirements.txt
Django==2.2.19
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-tinymce==3.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
flake8==3.9.2
importlib-metadata==4.5.0
mccabe==0.6.1
Pillow==8.2.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
python-dotenv==0.19.0
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
zipp==3.4.1
